Add a class that implements a Gaussian state
Source intensities are often described by a Gaussian in MEM methods. The
new mem.network.variables.GaussianState class represents the
distribution of instensities that can be described by a mean and a
variance.

The class has two methods: partition and project. The partition method
evaluates the partition function of the state for a given value of the
Lagrange multipliers. In addition to the value, it also returns the
Jacobian of the partition function. The project methods projects the
state into another space. Its intended use is to project the state from
the source space to the sensor space by using the lead field matrix as a
projector.