Add classes that represent the main variable types
In MEM methods, the Bayesian network contains two types of hidden
variables: clusters and connections.

The mem.network.variables.Cluster class represents a group of sources
with a shared state. For now, the functionality of this class is very
limited. It serves mainly as a container that enumerates the different
states achievable by a group of sources.

The mem.network.variables.Connection class represents a connection
between to clusters. It is empty for now and is included as a
placeholder.