Add a class for conditionnal probability tables
Variables of the Bayesian network of a MEM method can be in many
different states. The probability of being in a given state is kept in a
conditional probability table.

The mem.network.tables.ConditionalProbability class associates a
probability to every possible combination of variables states in its
domain. For example, if a table represents two variables having two
possible states each, the table has four entries.