Add a class for evidence tables
Evidence tables are probability tables that can be updated by evidence.
They can depend only on a single variable.

To prevent code repetition, a new BaseTable class was added. This new
class implements the functionality which is common to both the
ConditionalProbability and Evidence classes of the mem.network.tables
module.