Add a class for the Bayesian network
The Bayesian network models the interaction between sources, clusters,
and connections. It allows a probability to be associated with all
possible combinations of states of clusters and connections.

For now it is simply a container that ensures variables in a table
are part of the network and appear in the same order in all tables.