Simplify ConditionalProbability by using BaseTable
The mem.network.tables.ConditionalProbability was simplified by making
it a subclass of BaseTable. The idea is that a base table represents a
list of probabilities associated with a list of variable states. On the
other hand, a conditional probability table allows more manipulation,
such as marginalization and table products.