Add a class to compute marginals
To compute the entropy of the MEM Bayesian network, it is necessary to
compute marginals of conditional probability tables.

A marginal is constructed by using another table and a variable to
marginalize. This allows information on the source of the resulting
marginal to be kept. It is expected that updates to the marginal will be
faster using this approach rather than implementing marginals as a
method.