Add a class to compute table products
To compute the entropy of the MEM Bayesian network, it is necessary to
compute products of probability tables.

A table product is constructed using two tables and the list of
variables in the world. The list is necessary because the product relies
on the order of the variables. When computing a product, a reference to
the source tables is stored. This should allow efficient updates to the
product.