Add convenience class for zero one clusters
Cluster whose mean and variance are not truly important are created for
testing purpose. This creates a lot of redundant code.

A new convenience class is added to quickly create clusters with two
states having a mean of 0 and 1 and a variance of 0. This new class is
used to simplify many tests.