Add log_partition method to the network
The log partition of a network is one term in the entropy of a MEM
solution. It represents the likelyhood of an arrangement of sources.

The log_partition method of the network class evaluates the log
partition at a specific value of the Lagrange multipliers. The current
implementation is naive and will need to be improved.